Update Repository value from Integer to Long and fix docx 12-18

+++ dev/backend/haomibo/src/main/java/com/nuctech/ecuritycheckitem/repositories/FromConfigIdRepository.java
@@ -17,8 +17,8 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface FromConfigIdRepository extends
-        JpaRepository<FromConfigId, Integer>,
-        PagingAndSortingRepository<FromConfigId, Integer>,
+        JpaRepository<FromConfigId, Long>,
+        PagingAndSortingRepository<FromConfigId, Long>,
         QuerydslPredicateExecutor<FromConfigId> {
 
 }
