+++ dev/backend/haomibo/src/main/java/com/nuctech/ecuritycheckitem/repositories/SysDataGroupRepository.java
@@ -17,8 +17,8 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface SysDataGroupRepository extends
-        JpaRepository<SysDataGroup, Integer>,
-        PagingAndSortingRepository<SysDataGroup, Integer>,
+        JpaRepository<SysDataGroup, Long>,
+        PagingAndSortingRepository<SysDataGroup, Long>,
         QuerydslPredicateExecutor<SysDataGroup> {
 
 }
